[Fixes #66] Reads could encounter partially completed writes.
Approach:

* write files to a random tmp file
* once complete, use renameFile to atomically move the tmp file to the
  expected outputFile

Performance impact:

In general, at-least when being used as part of broccoli's build system,
this library is most sensitive to read performance and this approach
makes no changes to that flow. As for the right flow, testing suggests
file rename to be relatively small overhead relative to actually writing
the file.

Future work:

* when in same process: reads may want to wait for corresponding writes to
  finish before reading.
* we may want to partition files further to prevent single directories
  with a huge number of contained files.